store index.js & gameslice  converted to typescript
